🚚 Move CcForExp to another file
